feat : scaffold new database

+++ projet_csharp_travel_plan/projet_csharp_travel_plan/Models/ActiviteCategorie.cs
@@ -5,7 +5,7 @@ namespace projet_csharp_travel_plan.Models;
 
 public partial class ActiviteCategorie
 {
-    public int IdCatActiv { get; set; }
+    public short IdCatActiv { get; set; }
 
     public string Nom { get; set; } = null!;
