+++ projet_csharp_travel_plan/projet_csharp_travel_plan/Models/LogementCategorie.cs
@@ -5,7 +5,7 @@ namespace projet_csharp_travel_plan.Models;
 
 public partial class LogementCategorie
 {
-    public int IdLogementCategorie { get; set; }
+    public short IdLogementCategorie { get; set; }
 
     public string Nom { get; set; } = null!;
