+++ projet_csharp_travel_plan/projet_csharp_travel_plan/Models/TransportCategorie.cs
@@ -5,7 +5,7 @@ namespace projet_csharp_travel_plan.Models;
 
 public partial class TransportCategorie
 {
-    public int IdCategorieTransport { get; set; }
+    public short IdCategorieTransport { get; set; }
 
     public string Nom { get; set; } = null!;
